1 part access

+++ libft.h
@@ -6,7 +6,7 @@
 /*   By: mdark <mdark@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/10/14 17:04:11 by mdark             #+#    #+#             */
-/*   Updated: 2021/10/19 13:38:11 by mdark            ###   ########.fr       */
+/*   Updated: 2021/10/19 14:35:55 by mdark            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -41,7 +41,7 @@ char	*ft_strchr(char *str, int num);
 char	*ft_strdup(const char *str);
 size_t	ft_strlcat(char *dst, const char *src, size_t size);
 size_t	ft_strlcpy(char *dst, const char *str, size_t size);
-int		ft_strlen(const char *str);
+size_t	ft_strlen(const char *str);
 int		ft_strncmp(const char *s1, const char *s2, size_t n);
 char	*ft_strnstr(const char *str1, const char *str2, size_t len);
 char	*ft_strrchr(char *str, int ch);
